Code transitioning from PortableContacts.php to GServer.php
