+++ src/Model/Contact.php
@@ -1481,7 +1481,7 @@ class Contact extends BaseObject
 			$data = Probe::uri($url, "", $uid);
 			// Ensure that there is a gserver entry
 			if (!empty($data['baseurl']) && ($data['network'] != Protocol::PHANTOM)) {
-				PortableContact::checkServer($data['baseurl']);
+				GServer::check($data['baseurl']);
 			}
 		}
