+++ src/Worker/DiscoverPoCo.php
@@ -12,6 +12,7 @@ use Friendica\Core\Worker;
 use Friendica\Database\DBA;
 use Friendica\Model\GContact;
 use Friendica\Model\Contact;
+use Friendica\Model\GServer;
 use Friendica\Network\Probe;
 use Friendica\Protocol\PortableContact;
 use Friendica\Util\DateTimeFormat;
@@ -86,7 +87,7 @@ class DiscoverPoCo
 				return;
 			}
 			$result = "Checking server ".$server_url." - ";
-			$ret = PortableContact::checkServer($server_url);
+			$ret = GServer::check($server_url);
 			if ($ret) {
 				$result .= "success";
 			} else {
@@ -186,7 +187,7 @@ class DiscoverPoCo
 				$server_url = $user["server_url"];
 			}
 
-			if ((($server_url == "") && ($user["network"] == Protocol::FEED)) || $force_update || PortableContact::checkServer($server_url, $user["network"])) {
+			if ((($server_url == "") && ($user["network"] == Protocol::FEED)) || $force_update || GServer::check($server_url)) {
 				Logger::log('Check profile '.$user["url"]);
 				Worker::add(PRIORITY_LOW, "DiscoverPoCo", "check_profile", $user["url"]);
 
@@ -237,7 +238,7 @@ class DiscoverPoCo
 
 				$server_url = Contact::getBasepath($jj->url);
 				if ($server_url != '') {
-					if (!PortableContact::checkServer($server_url)) {
+					if (!GServer::check($server_url)) {
 						Logger::log("Friendica server ".$server_url." doesn't answer.", Logger::DEBUG);
 						continue;
 					}
